Stokhos:  Get SFS specialization of MP::Vector working again.
Mostly this is copying various things I have added to the main
MP::Vector implementation to the SFS specialization.  There is now a
define in Stokhos_ConfigDefs.h that determines whether it is used
(default is off).  I also redid how the overloaded operators are defined
so that the relevant pragmas can be used.  If you turn it on, everything
in my build appears to work.

(cherry picked from commit 467e70431ddb01232df521754c3187db3cced765)